+++ src/task-runtime.ts
@@ -7,6 +7,7 @@ import { gray, underline } from "chalk";
 import { PROJEN_DIR } from "./common";
 import * as logging from "./logging";
 import { TasksManifest, TaskSpec, TaskStep } from "./task-model";
+import { makeCrossPlatform } from "./util/tasks";
 
 const ENV_TRIM_LEN = 20;
 const ARGS_MARKER = "$@";
@@ -376,29 +377,4 @@ interface ShellOptions {
   /** @default false */
   readonly quiet?: boolean;
   readonly extraEnv?: { [name: string]: string | undefined };
-}
-
-/**
- * Makes a cross-shell command that works on both Windows and Unix-like systems.
- *
- * @param command The command to make cross-platform.
- * @returns
- */
-export function makeCrossPlatform(command: string): string {
-  const isWindows = process.platform === "win32";
-  if (!isWindows) {
-    return command;
-  }
-
-  return command
-    .split("&&")
-    .map((subcommand) => {
-      const trimmedSubcommand = subcommand.trim();
-      const cmd = trimmedSubcommand.split(" ")[0];
-
-      const supportedByShx = ["cat", "cp", "mkdir", "mv", "rm"].includes(cmd);
-
-      return supportedByShx ? `shx ${trimmedSubcommand}` : trimmedSubcommand;
-    })
-    .join(" && ");
 }
